missing cast?

+++ datasets/octhdf5.py
@@ -46,7 +46,7 @@ class OCTHDF5Dataset(Dataset):
 
 
         image = self.image_set[idx]
-        label = self.label_set[idx]
+        label = self.label_set[idx].astype(np.float32)
         # image = np.expand_dims(image, axis=-1)
         image = np.concatenate([image, image, image], axis=-1)
